% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allReg}
\alias{allReg}
\title{All Subsets Regression}
\usage{
allReg(x, y, wt = rep(1, nrow(x)), nmax = ncol(x), nbst = 3,
  na.rm.x = TRUE, lin.dep = 10)
}
\arguments{
\item{x}{matrix of candidate explanatory variables.}

\item{y}{the response variable.}

\item{wt}{the weight variable if needed.}

\item{nmax}{the maximum number of explanatory variables to include in the
largest model.}

\item{nbst}{the number of best models to determine for each subset size.}

\item{na.rm.x}{logical, if \code{TRUE}, then rows with missing values in
\code{x} and the corresponding elements in \code{y} will be removed prior
to analysis. If \code{FALSE}, then missing values are not removed and the
analysis will fail if there are missing values in \code{x}.}

\item{lin.dep}{a value to protect against linear dependencies; the number of
the number of observations must be greater than the number of columns in
\code{x} plus \code{lin.dep}.}
}
\value{
A data frame containing these columns: \item{model.formula}{the
subset model formula} \item{nvars}{the size (number of variables in the
subset model} \item{stderr}{the standard error of the subsbet model}
\item{R2}{the coefficient of determination for the subset model}
\item{adjr2}{the adjusted r-squared of the subbset model} \item{Cp}{Mallow's
Cp for the subset model} \item{press}{the press statistic for the subset
model}
}
\description{
Creates a table of the best subsets of explanatory variables for a response
variable.
}
\note{
This function is a wrapper for the \code{regsubsets} function in the
leaps package.
}
\examples{
# See the regression vignette for examples
.pager <- options("pager")
options(pager="console")
vignette(package="smwrStats")
options(.pager)
}
\references{
Helsel, D.R. and Hirsch, R.M., 2002, Statistical methods in
water resources: U.S. Geological Survey Techniques of Water-Resources
Investigations, book 4, chap. A3, 522 p.\cr

Mallow, C.L., 1973, Come comments of Cp: Technometrics, v. 15, p.
661--675.\cr

Miller, A.J., 1990, Subset selection in regression in Monographs on
Statistics and Applied Probability 40: London, Chapman and Hall.\cr
}
\seealso{
\code{\link{regsubsets}}
}
\keyword{models}
\keyword{regression}

