% Generated by roxygen2 (4.0.2): do not edit by hand
\name{confirm.seasonalPeak}
\alias{confirm.seasonalPeak}
\title{Confirm Seasonal Peak}
\usage{
\method{confirm}{seasonalPeak}(x, all = FALSE, plot.only = FALSE, ...)
}
\arguments{
\item{x}{a seasonalPeak object.}

\item{all}{a logical value indicating whether to accept the peak without
interactive user input or to force the user to process the peak. The default
value is \code{FALSE}, forcing the user to process the peak. Can also be set
to either 1 or 2, indicating the number of peaks.}

\item{plot.only}{a logical value indicating that only a plot is desired. If
\code{TRUE}, then \code{x} is returned invisibly and unchanged.}

\item{\dots}{not used, required for compatibility with other methods.}
}
\value{
An object of class seasonalPeak. The confirmed object is a single
value that represents the estimate of the timing of the peak and four or
five aadditional ttributes.\cr

NumPeaks: the number of seasonal peaks; either 1 or 2.\cr Models: candidate
loading models. The number indicates the number of months of constituent
loading.\cr hlife: candidate half-life values. The muner indicates the
half-life in terms of months. \cr Confirmed: logical indicating that the
object has not been confirmed.\cr

If \code{NumPeaks} is 2, then an additional attirbute \code{Second.peak}
that is a data frame of candidate parameters for the second peak is
included. See \code{\link{seasonalWave}} for details.
}
\description{
Reviews/accepts the results of an analysis: method for "seasonalPeak" object.
}
\examples{
\dontrun{
library(smwrData)
data(QW05078470)
# Simply click on the identified peak, and enter 1 for a single peak.
confirm(with(QW05078470, seasonalPeak(dectime(DATES), P00665)))
}
}
\references{
Vecchia, A.V., Martin, J.D., and Gilliom, R.J., 2008, Modeling
variability and trends in pesticide concentrations in streams: Journal of
the American Water Resources Association, v. 44, no. 5, p. 1308-1324.
}
\seealso{
\code{\link{seasonalWave}}, \code{\link{seasonalPeak}}
}
\keyword{manip}

