% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grubbs.test}
\alias{grubbs.test}
\title{Test for Outlier}
\usage{
grubbs.test(x, alternate = "two.sided")
}
\arguments{
\item{x}{a vector of numeric values. Missing values are allowed, but are
ignored in the calculation.}

\item{alternate}{the alternate hypothesis; must be "two.sided" for either a
high or a low outlier, "high" to test only for a high outlier, or "low" to
test only for a lopw outlier.}
}
\value{
An object of class "htest" having the following components:
\item{statistic}{ the value of the test statistic. } \item{p.value}{ the
attained p-value for the test. } \item{data.name}{ a character string
describing the name of the data used in the test. } \item{alternative}{ a
description of the altrnative and null hypotheses. } \item{method}{ a
description of the method. }
}
\description{
Test for a single outlier in a sample from a normal distribution.
}
\examples{
# A random sample with a high value
set.seed(100)
grubbstest <- rnorm(32)
grubbs.test(grubbstest)
qqnorm(grubbstest)
}
\references{
Grubbs, F., 1969, Procedures for Detecting Outlying Observations
in Samples, Technometrics, v. 11, no. 1, pp. 1-21.
}
\seealso{
\code{\link{pgrubbs}}, \code{\link{qgrubbs}}
}
\keyword{htest}

