% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jackknifeMove.1}
\alias{jackknifeMove.1}
\title{Jackknife move.1}
\usage{
jackknifeMove.1(object, newdata, type = c("response", "link"))
}
\arguments{
\item{object}{an object of class "move.1" on which to base the predicted
values.}

\item{newdata}{an optional data.frame in which to look for variables with
which to predict.  If omitted, then the calibration data are used; the response
values are sorted from smallest to largest and \code{type} is set to "link" if it
is not set to "response" in the call.}

\item{type}{the type of prediction ("response" or "link"). See
\bold{Details}.}
}
\value{
A vector of predictions matching \code{newdata} or the model data.
}
\description{
Computes the bias and variance of move.1 predictions using the jackknife estimator.
}
\details{
If \code{type} is "response," then the predicted values are
back-transformed. Otherwise, the predicted values are computed directly from
the model equation.
}
\examples{
library(smwrData)
data(IonBalance)
# Build model for non missing Alkalinity
IB.move <- move.1(Anion_sum ~ Cation_sum, data=IonBalance, subset=abs(Pct_Diff) < 10)
print(IB.move)
# Predict Anion_sum for missing Alkalinity
predict(IB.move, IonBalance[1, ])
}
\references{
Lorenz, D.L., 2015, smwrStats-an R package for analyzing hydrologic data,
version 0.7.0: U.S. Geological Survey Open-File Report 2015-XXXX, XX p.
}
\seealso{
\code{\link{move.1}}, \code{\link{predict.move.1}}
}
\keyword{models}
\keyword{regression}

