% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kensen.test}
\alias{kensen.test}
\title{Test for a Trend}
\usage{
kensen.test(y, t, n.min = 10)
}
\arguments{
\item{y}{the data collected over time. Missing values (NAs) are allowed and
removed before computations.}

\item{t}{the time corresponding to each observations in \code{y}. Missing
values are allowed only where \code{y} is missing. These should be expressed
as Julian or decimal time and must be strictly increasing.}

\item{n.min}{the minimum number of observations for adjusting the p-value
for serial correlation.  Used when \code{t} are uniformly spaced to adjust
the p-value for serial correlation. Any value larger than the number of
observations in \code{y} or \code{Inf} will suppress the adjustment.}
}
\value{
An object of class "htest" containing the following components:
\item{method}{ a description of the method.  } \item{statistic}{ the value
of Kendall's tau. } \item{p.value}{ the p-value. } \item{estimate}{ a named
vector containing the Sen estimate of the slope in units per year, the
median value of the data, the median value of time, the number of
observations, and if the serial correction is applied, the effective number
of observations (n*). } \item{data.name}{ a string containing the actual
name of the input series. } \item{coef}{ a vector of an estimate of the
intercept and slope. } \item{alternative}{ a character string describing
alternative to the test ("two.sided"). } \item{null.value}{ the value for
the hypothesized slope (0).  }
}
\description{
Tests for a temporal trend using Kendall's tau and computes the Sen slope
estimate of the trend.
}
\note{
A straight line of the form \cr \preformatted{ ytrend = sen.slope * (
t - median.time ) + median.data} may be used as a trend line for graphically
portraying or detrending the data. It goes through the point\cr
\preformatted{ (t,y) = ( median.time , median.data )} with slope sen.slope.
\cr

Many tied values can cause misleading results.\cr

The p-values for uniformly spaced data (\code{t} values unit value like
years) are adjusted for lag-1 autoregressive serial correlation according to
the method described by Yue and Wang (2004) that adjusts for trend. In
keeping with the logic of \code{seaken}, the p-value adjustment is never
performed for fewer than 10 observations. The user can suppress the
adjustment by setting the value of \code{n.min} to \code{Inf}.
}
\examples{
\dontrun{
library(smwrData)
data(SaddlePeaks)
with(SaddlePeaks, kensen.test(Flow, Year))
}
}
\references{
Conover, W.J., 1980, Practical nonparametric statistics (2d
ed.): New York, Wiley, 512 p.\cr

Helsel, D.R., and Hirsch, R.M., 2002, Statistical methods in water
resources: U.S. Geological Survey Techniques of Water-Resources
Investigations, book 4, chap. A3, 522 p.\cr

Hirsch, R.M., Alexander, R.B. , and Smith, R.A., 1991, Selection of methods
for the detection and estimation of trends in water quality: Water Resources
Research, v. 27 p. 803--813.\cr

Kendall, M.G., 1938, A new measure of rank correlation: Biometrika v. 30, p.
81--89.\cr

Kendall, M.G., 1976, Rank correlation methods (4th ed.): London, Griffin,
202 p.\cr

Sen, P.K., 1968, Estimates of regression coefficient based on Kendall's tau:
Journal of the American Statisical Association, v. 63, p. 1379--1389.\cr

Yue, S. and Wang. C., 2004, The Mann-Kendall test modified by effective
sample size to detect trend in serially correlated hydrological series:
Water Resources Management v. 18, p. 201-218.
}
\seealso{
\code{\link{dectime}}, \code{\link{seaken}}
}
\keyword{htest}

