% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/percentile.R
\name{percentile}
\alias{percentile}
\alias{percentile.default}
\title{Empirical Cumulative Percent}
\usage{
percentile(x, q, test = ">=", na.rm = TRUE, percent = TRUE, ...)

\method{percentile}{default}(x, q, test = ">=", na.rm = TRUE,
  percent = TRUE, ...)
}
\arguments{
\item{x}{a numeric vector representing the observed data.}

\item{q}{a vector of quantiles for which the cumulative percent or percent
exceedence is desired.}

\item{test}{a character string indicating the test. The default value, '>=,'
is the percent equalling or exceeding the quantile and '<' would return the
cumulative percent below the quantile.}

\item{na.rm}{a logical value indication whether missing values (NAs) should
be removed or not. If na.rm is \code{FALSE} and there are missing values in
\code{x}, then the result will be NA.  The default value is \code{TRUE}.}

\item{percent}{a logical value indicating whether the result should be
expressed as a percent or proportion. The default value, \code{TRUE}, will
express the result as a percent.}

\item{\dots}{not used, required for method function}
}
\value{
A named vector as long as \code{q} corresponding to the requested
value.
}
\description{
The empirical cumulative percent or percent exceedence of observed data is
computed for specific values.\cr
}
\note{
The stats package contains the \code{ecdf} function that performs a
similar function when \code{test} is "<=."
}
\examples{
set.seed(2342)
Xr <- rlnorm(24)
# The percentage of the observarions greater than or equal to 2
percentile(Xr, 1:5)
}
\seealso{
\code{\link{ecdf}}
}
\keyword{manip}
\keyword{math}
\keyword{univar}

