% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ppcc.test}
\alias{ppcc.test}
\title{Test for Normality}
\usage{
ppcc.test(x)
}
\arguments{
\item{x}{a vector of numeric values. Missing values are allowed, but are
ignored in the calculation.}
}
\value{
An object of class "htest" having the following components:
\item{statistic}{ the value of the test statistic. } \item{p.value}{ the
attained p-value for the test. } \item{data.name}{ a character string
describing the name of the data used in the test. } \item{method}{ a
description of the method. }
}
\description{
Computes the probability plot correlation coefficient test for departures
from normality.
}
\note{
The PPCC test is attractive because it has a simple, graphical
interpretation: it is a measure of the correlation in a Q-normal plot of the
data.  As such, it is related to the Shapiro-Wilk test (Shapiro and Wilk,
1965) for normality.\cr

The distribution function of the test statistic is empirical. This
application uses the "pocket calculator" approximation for computing the
p-value of the observed statistic (Royston, 1992).\cr
}
\examples{
## These data should produce an attained p-value less than 0.001
set.seed(45)
ppcc.test.data <- rnorm(32)
qqnorm(ppcc.test.data)
abline(mean(ppcc.test.data), sd(ppcc.test.data))
ppcc.test(ppcc.test.data)
}
\references{
Filliben, 1975, The PPCC test for normality: Technometrics, v.
17, no. 1, p. 111--117.\cr

Looney, S.W., and Gulledge, T.R., 1985, Use of the correlation coefficient
with normal probability plots: The American Statistician, v. 39, p.
75--79.\cr

Royston, J.P., 1992, A pocket-calculator algorithm for the Shapiro-Francia
test of non-normality--an application to medicine: Statistics in Medicine,
v. 12, p. 181--184.\cr

Shapiro, S.S., and Wilk, M.B., 1965, An analysis of variance test for
normality (complete samples): Biometrika, v. 52, p. 591--611.
}
\seealso{
\code{\link{shapiro.test}}
}
\keyword{htest}

