% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.move.1}
\alias{predict.move.1}
\title{Predict Maintenance of Variance Extension, Type 1}
\usage{
\method{predict}{move.1}(object, newdata, type = c("response", "link"),
  var.fit = FALSE, ...)
}
\arguments{
\item{object}{an object of class "move.1" on which to base the predicted
values.}

\item{newdata}{an optional data.frame in which to look for variables with
which to predict.  If omitted, then the fitted values are used.}

\item{type}{the type of prediction ("response" or "link"). See
\bold{Details}.}

\item{var.fit}{logical if \code{TRUE}, then compute the variance of the predicted
values. If \code{FALSE}, then the varainces are not computed.}

\item{\dots}{not used, required for method function.}
}
\value{
If \code{var.fit} is \code{FALSE}, then a vector of predictions matching
\code{newdata} or the model data. If \code{var.fit} is \code{TRUE}, then a data frame
containing the columns:
\item{fit}{the predicted values}
\item{var.fit}{the variance of the predicted values}
}
\description{
Predicts new values from a maintenance of variance extension, type 1 (MOVE.1)
model fit.
}
\details{
If \code{type} is "response," then the predicted values are
back-transformed. Otherwise, the predicted values are computed directly from
the model equation.
}
\examples{
library(smwrData)
data(IonBalance)
# Build model for non missing Alkalinity
IB.move <- move.1(Anion_sum ~ Cation_sum, data=IonBalance, subset=abs(Pct_Diff) < 10)
print(IB.move)
# Predict Anion_sum for missing Alkalinity
predict(IB.move, IonBalance[1, ])
}
\references{
Lorenz, D.L., 2015, smwrStats-an R package for analyzing hydrologic data,
version 0.7.0: U.S. Geological Survey Open-File Report 2015-XXXX, XX p.
}
\seealso{
\code{\link{move.1}}, \code{\link{jackknifeMove.1}}
}
\keyword{models}
\keyword{regression}

