% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.move.2}
\alias{predict.move.2}
\title{Predict Maintenance of Variance Extension, Type 2}
\usage{
\method{predict}{move.2}(object, newdata, type = c("response", "link"))
}
\arguments{
\item{object}{an object of class "move.2" on which to base the predicted
values.}

\item{newdata}{an optional data.frame in which to look for variables with
which to predict.  If omitted, then the fitted values are used.}

\item{type}{the type of prediction ("response" or "link"). See
\bold{Details}.}

\item{\dots}{not used, required for method function.}
}
\value{
A vector of predictions matching \code{newdata} or the model data.
}
\description{
Predict new values from a maintenance of variance extension, type 2 (MOVE.2)
model fit.
}
\details{
If \code{type} is "response," then the predicted values are
back-transformed. Otherwise, the predicted values are computed directly from
the model equation.
}
\note{
If lag was set to a non-zero value in the call to \code{move.2}, then the
explanatory variable is lagged only when predictiong values from the calibration
data (\code{newdata} is not supplied.) This facilitates prediction of selected
statistics at the response site rather than the complete record.
}
\examples{
\dontrun{
# See the vignette:
vignette("RecordExtension", package="smwrStats")
}
}
\seealso{
\code{\link{move.2}}
}
\keyword{models}
\keyword{regression}

