% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.cor.all.R
\name{print.cor.all}
\alias{print.cor.all}
\title{Print Objects}
\usage{
\method{print}{cor.all}(x, digits = 4, lower = TRUE, ...)
}
\arguments{
\item{x}{an object of class "cor.all" from \code{cor.all}.}

\item{digits}{the number of significant digits to print numeric data.}

\item{lower}{logical, print only the lower triangular matrix? Otherwise,
print the full, square matrix.}

\item{\dots}{not used for method, required for other methods.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Print the results of a correlation test (\code{cor.all}).
}
\note{
The printed output contains a description of the test; and 3 lines for
each comparison that is printed: the correlation statistic, the attained
p-value, and the number of pairs.
}
\seealso{
\code{\link{cor.all}}, \code{\link{printCor}}
}

