\name{print.move.1}
\alias{print.move.1}
\title{
Print Objects
}
\description{
Print the results of a move.1 analysis (\code{move.1}).
}
\usage{
\method{print}{move.1}(x, digits = 4, ...)
}
\arguments{
  \item{x}{
an object of class "move.1" from \code{move.1}.
}
  \item{digits}{
the number of significant digits to print numeric data.
}
  \item{\dots}{
additonal arguments for printing numeric values.
}
}
\value{
The object \code{x} is returned invisibly.
}

\note{
The printed output contains the call, the regression coefficients, and selected statistics
of the variables.
}

