% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.move.2.R
\name{print.move.2}
\alias{print.move.2}
\title{Print Objects}
\usage{
\method{print}{move.2}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class "move.2" from \code{move.2}.}

\item{digits}{the number of significant digits to print numeric data.}

\item{\dots}{additonal arguments for printing numeric values.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Print the results of a MOVE.2 analysis (\code{move.2}).
}
\note{
The printed output contains the call, the regression coefficients, and
selected statistics of the variables.
}

