% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.multReg.R
\name{print.multReg}
\alias{print.multReg}
\title{Print Objects}
\usage{
\method{print}{multReg}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class "multReg" from \code{multReg}.}

\item{digits}{the number of significant digits to print numeric data.}

\item{\dots}{not used for method, required for other methods.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Print the results of a multiple regression diagnostic analysis
(\code{multReg}).
}
\note{
The printed output contains the regression model call; a summary of
the residuals; A table of the coefficients with their estiamtes, standard
errors, t vlaues, and attained probability levels; the residual standard
error; R-squared and F-statistical summaries; Model comparison statistics;
if more than one explanatory variable a type II sum-of-squares analysis of
variance table and variance inflation factors; and selected test criteria
with observations that exceed one of more of the criteria.
}

