% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.optimBoxCox}
\alias{print.optimBoxCox}
\title{Print Object}
\usage{
\method{print}{optimBoxCox}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class "optimBoxCox" from \code{optimBoxCox}.}

\item{digits}{the number of significant digits to print numeric data.}

\item{\dots}{not used for method, required for other methods.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Prints the results of a multivariate unconditional Box-Cox transformation.
}
\note{
The printed output contains a table showing the power transformation
values and their standard errors.
}
\seealso{
\code{\link[smwrBase]{boxCox}}
}

