% Generated by roxygen2 (4.0.2): do not edit by hand
\name{regken}
\alias{regken}
\title{Trend Test}
\usage{
regken(series, correct = nrow(series) > 9)
}
\arguments{
\item{series}{a numeric matrix with rows representing the annual observations and
columns representing the sites. Missing values are permitted.}

\item{correct}{logical, if \code{TRUE}, then apply the correction for
cross correlation among sites. if \code{FALSE}, then do not apply the correction.}
}
\value{
An object of class "htest" also inhereting class "seaken" containing
the following components:
\item{method}{ a description of the method. }
\item{statistic}{ the value of Kendall's tau. }
\item{p.value}{ the p-value. See \bold{Note}. }
\item{p.value.raw}{ the p-value computed without correction for cross
correlation. See \bold{Note}. }
\item{p.value.corrected}{ the p-value computed with correction for cross
correlation. See \bold{Note}. }
\item{estimate}{ a named vector containing the Sen estimate of the slope in
units per year, the median value of the data, and the median value of time.}
\item{data.name}{ a string containing the actual name of the input series
with the number of years and sites }
\item{alternative}{ a character string describing alternative to the test
("two.sided"). }
\item{null.value}{ the value for the hypothesized slope (0).}
\item{nyears}{ the number of years.}
\item{nseasons}{ the number of sites}
\item{series}{ the data that was analyzed.}
\item{seasonnames}{ the names of the sites}
}
\description{
Computes the regional Kendall trend test with Sen slope estimator.
}
\note{
The value of \code{p.value} is \code{p.value.raw} if there are fewer
than 10 years of data and is \code{p.value.corrected} otherwise.

The regional Kendall is described in Douglas and others (2000) and Helsel and others (2006).
The adjustment for spatial correlation used in \code{regken} is based on equation 13 in
Douglas and others (2000) and uses the Spearman correlation to account for monotonic correlations.
}
\examples{
# Need example?
}
\references{
Douglas, E.M., Vogel, R.M., and Kroll, C.N., 2000, Trends in floods and low flows in
the United States: impact of spatial correlation: Journal of Hydrology, v. 240, p. 90--105.

Helsel, D.R., Mueller, D.K., and Slack, J.R., 2006, Computer program for the Kendall
family of trend tests: U.S. Geological Survey Scientific Investigations Report 2005--5275, 4 p.
}
\seealso{
\code{\link{seaken}}
}
\keyword{htest}

