% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seasonalPeak.R
\name{seasonalPeak}
\alias{seasonalPeak}
\title{Seasonal Peak Timing}
\usage{
seasonalPeak(x, y)
}
\arguments{
\item{x}{a vector of decimal time representing dates and times. Missing
values are permitted and are removed before analysis.}

\item{y}{a vector of the data for which the peak is needed. Missing values
are permitted and are removed before analysis.}
}
\value{
An object of class seasonalPeak. The unconfirmed object is a single
value that represents the estimate of the timing of the peak and five
additional attributes.\cr

Data: a list of the \code{x} and \code{y} values where x is the fractional
part of the original decimal time data. Missing values have been removed.\cr
Smooth: a list of the \code{x} and \code{y} smoothed values. \cr Points: a
list of 361 evenly spaced xout and yout values. \cr Extra: pointers to all
the peaks in \code{Points}. \cr Confirmed: logical indicating that the
object has not been confirmed.
}
\description{
Compute the timing of the seasonal peak value. The timing of the seasonal
peak is needed for the seasonalWave model (Vecchia and others, 2008).
}
\details{
The timing of the peak of the data is computed by identifying the largest
value produced by smoothing that data with \code{supsmu}. The remaining data
in the attributes are used by using the seasonalPeak method of confirm.
}
\note{
The generic functions print and confirm have methods for object of
class seasonalPeak.
}
\examples{
library(smwrData)
data(QW05078470)
with(QW05078470, seasonalPeak(dectime(DATES), P00665))
## Should be:
# Default value: 0.499
# Alternate values: 0.497
}
\references{
Vecchia, A.V., Martin, J.D., and Gilliom, R.J., 2008, Modeling
variability and trends in pesticide concentrations in streams: Journal of
the American Water Resources Association, v. 44, no. 5, p. 1308-1324
}
\seealso{
\code{\link{seasonalWave}}, \code{\link{confirm.seasonalPeak}},
\code{\link{supsmu}}, \code{\link{print.seasonalPeak}}
}
\keyword{manip}

