% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/seriesPlot.seaken.R
\name{seriesPlot.seaken}
\alias{seriesPlot}
\alias{seriesPlot.seaken}
\title{Series Plot}
\usage{
\method{seriesPlot}{seaken}(x, SeasonLine = list(name = "", what = "vertical",
  color = "black"), SeasonPoint = list(name = "", what = "points", symbol =
  "circle", filled = TRUE, size = 0.09, color = "black"), yaxis.log = FALSE,
  yaxis.range = c(NA, NA), ylabels = 7, xlabels, xtitle = "",
  ytitle = deparse(substitute(x)), caption = "", margin = c(NA, NA, NA,
  NA), SeasonTrend = list(name = "", what = "lines", color = "green"),
  TrendLine = list(name = "", what = "lines", color = "blue"), ...)
}
\arguments{
\item{x}{data that can be treated as a regularly-spaced time series.}

\item{SeasonLine}{control parameters of the lines in the plot. See
\bold{Details}.}

\item{SeasonPoint}{control parameters of the points in the plot. See
\bold{Details}.}

\item{yaxis.log}{log-transform the y axis?}

\item{yaxis.range}{set the range of the y axis.}

\item{ylabels}{set the y-axis labels. See \code{\link{linearPretty}} for
details.}

\item{xlabels}{set the x-axis labels, may be a single numeric value
indicating the number of season in \code{x}. See \code{\link{namePretty}}
for details.}

\item{xtitle}{the x-axis title.}

\item{ytitle}{the y-axis title.}

\item{caption}{the figure caption.}

\item{margin}{the parameters of the margin of the plot area.}

\item{SeasonTrend}{control parameters of the trend line for each season. See
\bold{Details}.}

\item{TrendLine}{control parameters of the overall trend line compute by
\code{seaken}. See \bold{Details}.}

\item{\dots}{any additional arguments required for specific methods.}
}
\value{
Information about the graph.
}
\description{
Produce a plot of time-series data by season that shows seasonal and overall
trends.
}
\details{
The argument \code{what} for \code{SeasonLine} must be either "lines" or
"vertical."  See \code{\link{monthplot}} for more information.\cr The
argument \code{what} for either \code{SeasonPoint}, \code{SeasonTrend}, or
\code{trendLine} may be set to "none" to suppress drawing of that feature.
}
\seealso{
\code{\link{monthplot}}, \code{\link{seasonPlot}},
\code{\link{seaken}}
}
\keyword{hplot}

