% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{smwrStats-package}
\alias{smwrStats}
\alias{smwrStats-package}
\title{General tools for hydrologic data and trend analysis.}
\description{
Useful tools for the analysis of hydrologic data, not inlcuding censored
water-quality data.
}
\details{
\tabular{ll}{
Package: \tab smwrStats\cr
Type: \tab Package\cr
Version: \tab 0.7.5\cr
Date: \tab 2016-01-21\cr
License: \tab CC0\cr
}
Regression applications:\cr allReg\cr binaryReg\cr hosmerLemeshow.test\cr
leCessie.test\cr multReg\cr press\cr rmse\cr selBestWave\cr
roc\cr vif\cr predictDuan\cr predictFerguson\cr predictMVUE\cr
senSlope\cr seasonalPeak\cr seasonalWave\cr

Record extension applications:\cr move.1\cr jackknifeMove.1\cr move.2\cr
optimBoxCox\cr

Trend applications:\cr curvi\cr kensen.test\cr regken\cr seaken\cr
trends\cr serial.test\cr regken\cr trends\cr

Summary statistics:\cr cor.all\cr printCor\cr percentile\cr
quantile.numeric\cr qtiles.CI\cr skew\cr sumStats\cr timeWeightedMean\cr
multicomp.test\cr

Hypothesis tests:\cr grubbs.test\cr ppcc.test\cr serial.test\cr
}
\author{
Dave Lorenz <lorenz@usgs.gov>\cr

Maintainer: Dave Lorenz <lorenz@usgs.gov>
}
\references{
Helsel, D.R. and Hirsch, R.M., 2002, Statistical methods in
water resources: U.S. Geological Survey Techniques of Water-Resources
Investigations, book 4, chap. A3, 522 p.\cr

Lorenz, D.L., in preparation. smwrStats---an R package for the analysis of
hydrologic data, version 0.7.5.
}
\keyword{package}

