% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{USGSwsStats-package}
\alias{USGSwsStats}
\alias{USGSwsStats-package}
\title{General tools for hydrologic data and trend analysis.}
\description{
Useful tools for the analysis of hydrologic data, not inlcuding censored
water-quality data.
}
\details{
\tabular{ll}{ Package: \tab USGSwsStats\cr Type: \tab Package\cr Version:
\tab 0.6\cr Date: \tab 2013-08-14\cr License: \tab file LICENSE\cr }
Regression applications:\cr allReg\cr binaryReg\cr hosmerLemeshow.test\cr
leCessie.test\cr move.1\cr multicomp.test\cr multReg\cr press\cr rmse\cr
roc\cr vif\cr

Trend applications:\cr curvi\cr kensen.test\cr seaken\cr selBestWave\cr
senSlope\cr trends\cr

Summary statistics:\cr annualStats\cr cor.all\cr percentile\cr
quantile.numeric\cr qtiles.CI\cr skew\cr sumStats\cr

Hypothesis tests:\cr grubbs.test\cr ppcc.test\cr serial.test\cr
}
\author{
Dave Lorenz <lorenz@usgs.gov>\cr

Maintainer: Dave Lorenz <lorenz@usgs.gov>
}
\references{
Helsel, D.R. and Hirsch, R.M., 2002, Statistical methods in
water resources: U.S. Geological Survey Techniques of Water-Resources
Investigations, book 4, chap. A3, 522 p.\cr

Lorenz, D.L., in preparation. USGSwsStats--an R package for the analysis of
hydrologic data, version 0.6.
}
\keyword{package}

