% Generated by roxygen2 (4.0.2): do not edit by hand
\name{confirm}
\alias{confirm}
\alias{confirm.default}
\title{Confirm an Analysis}
\usage{
confirm(x, ...)

\method{confirm}{default}(x, ...)
}
\arguments{
\item{x}{the object to be confirmed}

\item{\dots}{additional arguments required for specific methods}
}
\value{
The object returned depends on the specific method.
}
\description{
Review and accept the results of an analysis
}
\note{
The default method simply returns the object and issues a warning.
}
\keyword{manip}

