% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grubbs.R
\name{qgrubbs}
\alias{Grubbs}
\alias{pgrubbs}
\alias{qgrubbs}
\title{The Grubbs Distribution}
\usage{
qgrubbs(alpha, N)

pgrubbs(G, N)
}
\arguments{
\item{alpha}{the significance level.}

\item{N}{the number of values in the sample.}

\item{G}{the maximum or minimum scaled difference from the mean.}
}
\value{
The function \code{pgrubbs} returns the attained p-value, not the
cumulative distribution, for the maximum scaled distance from the mean. the
funciton \code{qgrubbs} returns the one-sided critical value for the maximum
scaled difference from the mean.
}
\description{
The one-sided critical value and the attained signicance level for the
Grubbs test for an outlier in a sample from Normal distribution.
}
\examples{
# The difference is due to rounding errors
pgrubbs(c(.9, .95, .99), 32)
qgrubbs(c(5.905348, 5.483234, 5.159097), 32)
}
\references{
Grubbs, F., 1969, Procedures for Detecting Outlying Observations
in Samples, Technometrics, v. 11, no. 1, pp. 1-21.
}
\seealso{
\code{\link{grubbs.test}}
}
\keyword{distribution}

