% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multreg.R
\name{multReg}
\alias{multReg}
\title{Diagnostics for Linear Regression}
\usage{
multReg(object)
}
\arguments{
\item{object}{the linear regression model object}
}
\value{
A object of class "multReg" having the following components:\cr

\item{aovtab}{ the analysis of variance table, using the type II sum of
squares. See \bold{Note}. } \item{parmests}{ a summary of \code{object}.  }
\item{vif}{ a named vector of variance inflation factors. }
\item{diagstats}{ a data.frame containing the observed values, predicted
values, residuals, standardized residuals, studentized residuals, leverage,
Cook's D, and dfits for each observation. } \item{crit.val}{ a named vector
of the critical values for leverage, Cook's D, and dfits.  See \bold{Note} }
\item{flagobs}{ a logical vector indicating which observations exceeded at
least one of the critical values. } \item{object}{ the \code{lm} object. }
\item{x}{ the model matrix of explanatory variables. }
}
\description{
Computes diagnostics for linear regression
}
\note{
The type II sum of squares are calculated according to the principle
of marginality, testing each term after all others, except ignoring the
term's higher-order relatives. This type sum of squares is useful for
assessing the overall marginal effect of each term in the model.\cr

The critical values for the test criteria are computed as: leverage,
\emph{3p/n}; Cook's D, median quantile for the \emph{F} distribution with
\emph{p+1} and \emph{n-p} degrees of freedom; and dfits, the .01 quantile
of the \emph{grubbs} distribution for \emph{n} observations time the square
root of (p/n), where \emph{p}
is the number of parameters estimated in the regression and \emph{n} is the
number of observations.

Objects of class "multReg" have \code{print} and
\code{plot} methods.
}
\references{
Draper, N.R. and Smith, H., 1998, Applied Regression Analysis,
(3rd ed.): New York, Wiley, 724 p.\cr
}
\seealso{
\code{\link{lm}}, \code{\link{plot.multReg}},
}
\keyword{models}
\keyword{regression}

