% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optimBoxCox.R
\name{optimBoxCox}
\alias{optimBoxCox}
\title{Multivariate Unconditional Box-Cox Transformations}
\usage{
optimBoxCox(X, start = NULL)
}
\arguments{
\item{X}{a data frame or matrix of the data to find the optimized Box-Cox
transforms to produce multivariate normality. Can also be a numeric vector for
a simple Box-Cox transform to normality.}

\item{start}{a numeric vector of length matching the number of columns in \code{X}
to provide starting values for the Box-Cox transforms.}
}
\value{
An object of class "optimBoxCox" having these components:
\item{start}{ the starting values for the Box-Cox transformations.}
\item{criterion}{ the log-likelihood of the Box-Cox transformations.}
\item{names}{ the names of the columns.}
\item{lambda}{ the values of the Box-Cox transformations.}
\item{stderr}{ the standard errors of the Box-Cox transformations.}
\item{return.code}{ the convergence value returned by \code{optim}.}
\item{gm}{ the geometric means of the data in \code{X}.}
\item{data}{ the data in \code{X} with missing values removed.}
}
\description{
Compute Box-Cox transformations that maximimize the log likelihood
of the transformations.
}
\note{
The maximum likelihood estimate of the Box-Cox transformations corresponds to the
minimum determinant of the variance-covariance matrix of the transformed \code{X}. The
methodology is described in Andrews and others (1971).
}
\references{
Andrews, D.F., Gnanadesikan, R., and Warner, J.L., 1971, Transformations
of multivariate data: Biometrics, v. 27, p. 825--840.
}
\seealso{
\code{\link[smwrBase]{boxCox}}, \code{\link[stats]{optim}}
}

