% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.roc.R
\name{print.roc}
\alias{print.roc}
\title{Print Object}
\usage{
\method{print}{roc}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class "roc" from \code{roc}.}

\item{digits}{the number of significant digits to print numeric data.}

\item{\dots}{not used for method, required for other methods.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
PRint the results of a receiver operator characteristics (ROC) for a logistic
regression model.
}
\note{
The printed output contains the area under to ROC curve.
}

