% Generated by roxygen2 (4.0.2): do not edit by hand
\name{printCor}
\alias{printCor}
\title{Print Objects}
\usage{
printCor(x, criterion = 0.75)
}
\arguments{
\item{x}{an object of class "cor.all" from \code{cor.all} or the output from \code{cor}.}

\item{criterion}{a numeric value indicating the test criterion for showing a value.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Prints the results of a correlation test (\code{cor.all}) or correlation matrix
in a compact form indicating positive or negative correlation.
}
\note{
The printed output is a compressed table showing "+" where the value in \code{x}
is greater than \code{criterion}, "-" where the value in \code{x} is less than
-\code{criterion}, "\" on the diagonal, if \code{x} is symmetric, and "." where \code{x}
has a missing value, and " " otherwise.
}
\examples{
\dontrun{
library(smwrData)
data(TNLoads)
printCor(cor(TNLoads, method="spearman"), .5)
}
}
\seealso{
\code{\link{cor.all}}
}

