% Generated by roxygen2 (4.0.2): do not edit by hand
\name{roc}
\alias{roc}
\title{Receiver Operator Characteristics for Logistic Regression}
\usage{
roc(object)
}
\arguments{
\item{object}{an object of class "glm."}
}
\value{
An object of class "roc" haveing these components: \item{c.val}{the
area under the ROC curve} \item{table}{a 3-column matrix of the fitted
values, sensitivity and specificity}
}
\description{
Compute the receiver operator characteristics (ROC) for a logistic
regression model.
}
\note{
Objects of class "roc" have \code{print} and \code{plot} methods.
}
\references{
Gonen, M., 2006, Receiver Operating Characteristics (ROC)
Curves: SAS Users Group International Conference, Paper 210-31 18 p.,
available at http://www2.sas.com/proceedings/sugi31/210-31.pdf, last
accessed Octocer 26, 2011.
}
\keyword{regression}

