% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smwrStats-package.R
\docType{package}
\name{smwrStats-package}
\alias{smwrStats}
\alias{smwrStats-package}
\title{General tools for hydrologic data and trend analysis.}
\description{
Useful tools for the analysis of hydrologic data, not inlcuding censored
water-quality data.
}
\details{
\tabular{ll}{ Package: \tab smwrStats\cr Type: \tab Package\cr Version:
\tab 0.7.1\cr Date: \tab 2015-04-29\cr License: \tab file LICENSE\cr }
Regression applications:\cr allReg\cr binaryReg\cr hosmerLemeshow.test\cr
leCessie.test\cr multicomp.test\cr multReg\cr press\cr rmse\cr
roc\cr vif\cr

Record extension applications:\cr move.1\cr move.2\cr

Trend applications:\cr curvi\cr kensen.test\cr seaken\cr selBestWave\cr
senSlope\cr trends\cr

Summary statistics:\cr annualStats\cr cor.all\cr percentile\cr
quantile.numeric\cr qtiles.CI\cr skew\cr sumStats\cr

Hypothesis tests:\cr grubbs.test\cr ppcc.test\cr serial.test\cr
}
\author{
Dave Lorenz <lorenz@usgs.gov>\cr

Maintainer: Dave Lorenz <lorenz@usgs.gov>
}
\references{
Helsel, D.R. and Hirsch, R.M., 2002, Statistical methods in
water resources: U.S. Geological Survey Techniques of Water-Resources
Investigations, book 4, chap. A3, 522 p.\cr

Lorenz, D.L., in preparation. smwrStats---an R package for the analysis of
hydrologic data, version 0.7.1.
}
\keyword{package}

