% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.cor.all.R
\name{summary.cor.all}
\alias{summary.cor.all}
\title{Summarize Correlations}
\usage{
\method{summary}{cor.all}(object, p.adjust = "holm", variable = NULL, ...)
}
\arguments{
\item{object}{an object created by the \code{cor.all} function.}

\item{p.adjust}{a character string describing the method to use to adjust
the p-values to account for m ultiple comparisons. See
\code{\link{p.adjust}} for the options and details.}

\item{variable}{if the name of a variable, then summarize only those
correlations of this variable with all others, otherwise summarize all
combinations.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A data frame containing columns of the paired variables, the
correlation, the adjusted p-value, and the number of observations in the
correlation.
}
\description{
Extract a data frame that summarizes the correlations computed by
\code{cor.all}.
}
\examples{
library(smwrData)
data(TNLoads)
# Extract only the correlations with the log of total nitrogen
summary(cor.all(TNLoads[, 1:5]), variable="LOGTN")
}
\seealso{
\code{\link{cor.all}}
}
\keyword{htest}

